+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog for Windows-Installer-NPX
 <!-- ## **WORK IN PROGRESS**
 -->
-## **WORK IN PROGRESS**
-* Start/stop service by calling `iob.bat start/stop` 
+## 4.1.10 (2022-05-23)
+* Start/stop service by calling `iob.bat start/stop`
 
 ## 4.1.9 (2022-05-23)
 * Stop service before fix and then start service again